ContactDirectory Class working; requires methods saving user input to new database/file.
